+++ src/main/kotlin/com/terning/server/kotlin/ui/api/HomeController.kt
@@ -1,8 +1,5 @@
 package com.terning.server.kotlin.ui.api
 
-import com.terning.server.kotlin.application.internshipAnnouncement.InternshipAnnouncementService
-import com.terning.server.kotlin.application.internshipAnnouncement.dto.DetailAnnouncementResponse
-import com.terning.server.kotlin.application.internshipAnnouncement.dto.HomeAnnouncementsResponse
 import com.terning.server.kotlin.application.home.HomeService
 import com.terning.server.kotlin.application.home.dto.HomeResponse
 import com.terning.server.kotlin.application.home.dto.UpcomingDeadlineScrapResponse
@@ -11,7 +8,6 @@ import org.springframework.data.web.PageableDefault
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
@@ -20,9 +16,6 @@ import org.springframework.web.bind.annotation.RestController
 @RequestMapping("/api/v1/home")
 class HomeController(
     private val homeService: HomeService,
-@RequestMapping("/api/v1")
-class InternshipAnnouncementController(
-    private val internshipAnnouncementService: InternshipAnnouncementService,
 ) {
     @GetMapping("home")
     fun getInternshipAnnouncementsFilteredByUserFilter(
@@ -61,28 +54,6 @@ class InternshipAnnouncementController(
             ApiResponse.success(
                 status = HttpStatus.OK,
                 message = response.message,
-                result = response,
-            ),
-        )
-    }
-
-    @GetMapping("announcements/{internshipAnnouncementId}")
-    fun getDetailInternshipAnnouncement(
-        // TODO: @AuthenticationPrincipal userId: Long,
-        @PathVariable internshipAnnouncementId: Long,
-    ): ResponseEntity<ApiResponse<DetailAnnouncementResponse>> {
-        val userId: Long = 1 // TODO: @AuthenticationPrincipal 구현 시 제거
-
-        val response =
-            internshipAnnouncementService.getDetailAnnouncement(
-                userId = userId,
-                internshipAnnouncementId = internshipAnnouncementId,
-            )
-
-        return ResponseEntity.ok(
-            ApiResponse.success(
-                status = HttpStatus.OK,
-                message = "공고 상세 정보 불러오기에 성공했습니다",
                 result = response,
             ),
         )
